fix syntax, move web3 action creator to action folder and remove unnessary codes
