+++ .gitignore
@@ -21,5 +21,4 @@ yarn-debug.log*
 yarn-error.log*
 *.xml
 .idea
-src/env.json
 smartcontracts
