+++ src/components/ICO.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import { onModalShow } from '../actions/ModalAction';
 import ICOApp from './ICOApp';
 import ICOScan from './ICOScan';
-import { readSmartContract } from '../reducers/web3';
+import { readSmartContract } from '../actions/web3';
 import { isConnected } from '../utils/web3';
 import { errorMessage, resetRpc } from '../actions/ScanAction';
 
@@ -27,7 +27,6 @@ class ICO extends Component {
 
 
 const mapStateToProps = state => ({
-  showModal: state.modal.showModal,
   web3: state.modal.web3,
   isSmartContractLoaded: state.scan.isSmartContractLoaded,
 });
