+++ src/components/RPCProvider.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { web3Connection } from '../reducers/web3';
+import { web3Connection } from '../actions/web3';
 
 class RPCProvider extends Component {
