+++ src/containers/Scan.js
@@ -6,8 +6,7 @@ import ScanBoxDetails from '../components/ScanBoxDetails';
 import { default as config } from '../config.js';
 import { connect } from 'react-redux';
 import { Grid, Row, Col } from 'react-flexbox-grid';
-import { getLogs } from '../reducers/web3';
-import { readSmartContract } from '../reducers/web3';
+import { getLogs ,readSmartContract } from '../actions/web3';
 import { getNextICO } from '../utils';
 
 class Scan extends Component {
