+++ src/env.json
@@ -1,0 +1,3 @@
+{
+  "rpcHost" : "http://localhost:8545"
+}
